Fixed support for parallelism.
I was probably pretty darn tired while writing this code as I clearly
didn't realize the issue here. That being said: This code is still
pretty darn terrible due to the fact that we cannot just stop
mid-execution when we encounter an exception. Throwing a collective
exception is simply as good as it gets in this scenario.